+++ src/modules/kemix/kemix_mod.c
@@ -23,6 +23,7 @@
 #include <unistd.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 #include <time.h>
 
 #include "../../core/sr_module.h"
@@ -74,7 +75,7 @@ static sr_kemi_xval_t* ki_kx_get_ruri(sip_msg_t *msg)
 	}
 
 	if(msg->first_line.type == SIP_REPLY) {
-		/* REPLY doesnt have a ruri */
+		/* REPLY doesn't have a ruri */
 		sr_kemi_xval_null(&_sr_kemi_kx_xval, 0);
 		return &_sr_kemi_kx_xval;
 	}
@@ -107,7 +108,7 @@ static sr_kemi_xval_t* ki_kx_get_ouri(sip_msg_t *msg)
 	}
 
 	if(msg->first_line.type == SIP_REPLY) {
-		/* REPLY doesnt have a ruri */
+		/* REPLY doesn't have a ruri */
 		sr_kemi_xval_null(&_sr_kemi_kx_xval, 0);
 		return &_sr_kemi_kx_xval;
 	}
@@ -952,7 +953,7 @@ static sr_kemi_xval_t* ki_kx_get_nhuri(sip_msg_t *msg)
  */
 static int ki_kx_get_timestamp(sip_msg_t *msg)
 {
-	return (int)time(NULL);
+	return (int)(uint64_t)time(NULL);
 }
 
 /**
